pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ctn-testing"
version = "0.1.0"
description = "CTN kernel evaluation framework"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "pyyaml>=6.0",
    "scipy>=1.11",
    "numpy>=1.24",
    "tenacity>=9.1.2",
    "docile-benchmark>=0.3.4",
    "requests>=2.28",
    "streamlit>=1.29.0",
    "click>=8.0.0",
    "pydantic>=2.0",
]

[project.scripts]
ctn-test = "ctn_testing.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]
anthropic = [
    "anthropic>=0.40",
]
google = [
    "google-genai>=1.0",
]
openai = [
    "openai>=1.50",
]
all = [
    "anthropic>=0.40",
    "google-generativeai>=0.8",
    "openai>=1.50",
]
datasets = [
    "docile-benchmark>=0.3.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["ctn_testing*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.pyright]
include = ["ctn_testing", "scripts"]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
]
